+++ bgw-core/src/main/kotlin/tools/aqua/bgw/components/container/LinearLayout.kt
@@ -22,8 +22,8 @@ package tools.aqua.bgw.components.container
 import tools.aqua.bgw.components.gamecomponentviews.GameComponentView
 import tools.aqua.bgw.components.uicomponents.Orientation
 import tools.aqua.bgw.core.*
-import tools.aqua.bgw.observable.DoubleProperty
-import tools.aqua.bgw.observable.Property
+import tools.aqua.bgw.observable.properties.DoubleProperty
+import tools.aqua.bgw.observable.properties.Property
 import tools.aqua.bgw.visual.Visual
 
 /**
@@ -122,7 +122,7 @@ open class LinearLayout<T : GameComponentView>(
 		}
 	
 	init {
-		observableComponents.setInternalListenerAndInvoke {
+		observableComponents.setInternalListenerAndInvoke(listOf()) { _, _ ->
 			layout()
 		}
 		spacingProperty.internalListener = { _, _ -> layout() }
@@ -167,8 +167,8 @@ open class LinearLayout<T : GameComponentView>(
 
 	override fun T.onAdd() {
 		//add pos listeners
-		posXProperty.internalListener = { _, _ -> observableComponents.internalListener?.invoke() }
-		posYProperty.internalListener = { _, _ -> observableComponents.internalListener?.invoke() }
+		posXProperty.internalListener = { _, _ -> observableComponents.internalListener?.invoke(listOf(), listOf()) }
+		posYProperty.internalListener = { _, _ -> observableComponents.internalListener?.invoke(listOf(), listOf()) }
 	}
 
 	override fun T.onRemove() {