+++ bgw-core/src/main/kotlin/tools/aqua/bgw/components/layoutviews/GridPane.kt
@@ -115,11 +115,14 @@ open class GridPane<T : ComponentView>(
 		grid[columnIndex, rowIndex] = component?.apply {
 			this.widthProperty.internalListener = { _, _ -> updateGui?.invoke() }
 			this.heightProperty.internalListener = { _, _ -> updateGui?.invoke() }
-			this.posXProperty.internalListener = { _, _ -> updateGui?.invoke() }
-			this.posYProperty.internalListener = { _, _ -> updateGui?.invoke() }
+			this.posXProperty.internalListener = { _, _ -> posXProperty.setSilent(0.0) /*updateGui?.invoke()*/ }
+			this.posYProperty.internalListener = { _, _ -> posYProperty.setSilent(0.0) /*updateGui?.invoke()*/ }
 			this.parent = this@GridPane
 		}
-
+		
+		//TODO: Remove hard reset of position
+		component?.posXProperty?.setSilent(0.0)
+		component?.posYProperty?.setSilent(0.0)
 		updateGui?.invoke()
 	}
 	
@@ -510,14 +513,31 @@ open class GridPane<T : ComponentView>(
 	 */
 	override fun getChildPosition(child: ComponentView): Coordinate? =
 		grid.filter { it.component == child }.map {
-			val cols = renderedColWidths.toMutableList().subList(0, it.columnIndex)
-			val offsetX = cols.sum() + cols.size * spacing
-			
-			val rows = renderedRowHeights.toMutableList().subList(0, it.rowIndex)
-			val offsetY = rows.sum() + rows.size * spacing
-			
-			Coordinate(offsetX + child.posX, offsetY + child.posY)
+			getRelativeChildOffset(it) + Coordinate(xCoord = child.posX, yCoord = child.posY)
 		}.firstOrNull()
+	
+	/**
+	 * Function returning a contained child's coordinates within this [GridPane] relative to the top left corner with
+	 * scale.
+	 *
+	 * @param child Child to find.
+	 *
+	 * @return coordinate of given child in this [GridPane].
+	 */
+	override fun getActualChildPosition(child: ComponentView): Coordinate? =
+		grid.filter { it.component == child }.map {
+			getRelativeChildOffset(it) + Coordinate(xCoord = child.actualPosX, yCoord = child.actualPosY)
+		}.firstOrNull()
+	
+	private fun getRelativeChildOffset(it : GridIteratorElement<T>): Coordinate {
+		val cols = renderedColWidths.toMutableList().subList(0, it.columnIndex)
+		val rows = renderedRowHeights.toMutableList().subList(0, it.rowIndex)
+		
+		return Coordinate(
+			xCoord = cols.sum() + cols.size * spacing,
+			yCoord = rows.sum() + rows.size * spacing
+		)
+	}
 	
 	/**
 	 * Returns an [Iterator] over the grid components.