+++ bgw-core/src/main/kotlin/tools/aqua/bgw/components/layoutviews/Pane.kt
@@ -20,7 +20,10 @@
 package tools.aqua.bgw.components.layoutviews
 
 import tools.aqua.bgw.components.ComponentView
-import tools.aqua.bgw.observable.IObservable
+import tools.aqua.bgw.components.container.GameComponentContainer
+import tools.aqua.bgw.components.gamecomponentviews.GameComponentView
+import tools.aqua.bgw.observable.Observer
+import tools.aqua.bgw.observable.ValueObserver
 import tools.aqua.bgw.observable.ObservableArrayList
 import tools.aqua.bgw.util.Coordinate
 import tools.aqua.bgw.visual.Visual
@@ -45,7 +48,19 @@ open class Pane<T : ComponentView>(
 ) : LayoutView<T>(posX = posX, posY = posY, width = width, height = height, visual = visual), Iterable<T> {
     
     internal val observableComponents: ObservableArrayList<T> = ObservableArrayList()
-    
+
+    /**
+     * [onAdd] gets invoked anytime after a [ComponentView] is added to this
+     * [Pane] with the added [ComponentView] as its receiver.
+     */
+    var onAdd : (T.() -> Unit)? = null
+
+    /**
+     * [onRemove] gets invoked anytime after a [ComponentView] is removed from this
+     * [Pane] with the removed [ComponentView] as its receiver.
+     */
+    var onRemove : (T.() -> Unit)? = null
+
     /**
      * [ComponentView]s that are contained in this [Pane].
      */
@@ -54,16 +69,20 @@ open class Pane<T : ComponentView>(
         private set
     
     /**
-     * Adds a [listener] on the [observableComponents] list.
+     * Adds the [IValueObservable] to the [observableComponents] list.
+     *
+     * @param listener The [IValueObservable] to add.
      */
-    fun addComponentsListener(listener: IObservable) {
+    fun addComponentsListener(listener: ValueObserver<List<T>>) {
         observableComponents.addListener(listener)
     }
     
     /**
-     * Removes a [listener] from the [observableComponents] list.
+     * Removes the [ValueObserver] from the [observableComponents] list.
+     *
+     * @param listener The [ValueObserver] to remove.
      */
-    fun removeComponentsListener(listener: IObservable) {
+    fun removeComponentsListener(listener: ValueObserver<List<T>>) {
         observableComponents.removeListener(listener)
     }
     
@@ -94,8 +113,12 @@ open class Pane<T : ComponentView>(
         require(index in 0..observableComponents.size) {
             "Index $index is out of list range."
         }
-        
-        observableComponents.add(index, component.apply { parent = this@Pane })
+
+        observableComponents.add(index, component)
+        component.apply {
+            parent = this@Pane
+            onAdd?.invoke(this)
+        }
     }
     
     /**
@@ -132,30 +155,59 @@ open class Pane<T : ComponentView>(
             throw IllegalArgumentException(e.message)
         }
     }
-    
+
     /**
      * Removes the [ComponentView] specified by the parameter from this [Pane].
      *
      * @param component The [ComponentView] to remove.
+     *
+     * @return `true` if the [Pane] was altered by the call, `false` otherwise.
      */
     @Synchronized
-    fun remove(component: T) {
-        observableComponents.remove(component.apply { parent = null })
+    fun remove(component: T): Boolean {
+        if (observableComponents.remove(component)) {
+            component.parent = null
+            onRemove?.invoke(component)
+            return true
+        }
+        return false
     }
-    
+
     /**
      * Removes all [ComponentView]s from this [Pane].
      *
-     * @return [List] of all removed components.
+     * @return List of all removed components.
      */
     @Synchronized
-    fun removeAll(): List<T> {
+    fun clear(): List<T> {
         val tmp = observableComponents.toList()
-        observableComponents.forEach { it.parent = null }
-        observableComponents.clear()
+        tmp.map { remove(it) }
         return tmp
     }
-    
+
+    /**
+     * Removes all [ComponentView]s contained in [collection] from this [Pane].
+     *
+     * @param collection The [ComponentView]s to remove.
+     *
+     * @return `true` if the [Pane] was altered by the call, `false` otherwise.
+     */
+    @Synchronized
+    fun removeAll(collection: Collection<T>) : Boolean =
+        collection.map { remove(it) }.fold(false) { x,y -> x || y }
+
+    /**
+     * Removes all [ComponentView]s matching the [predicate] from this [Pane].
+     *
+     * @param predicate The predicate to evaluate.
+     *
+     * @return `true` if the [Pane] was altered by the call, `false` otherwise.
+     */
+    @Synchronized
+    fun removeAll(predicate: (T) -> Boolean): Boolean =
+        components.map { if (predicate(it)) remove(it) else false }.fold(false) { x, y -> x || y }
+
+
     /**
      * Returns the size of the components list.
      *
@@ -192,7 +244,18 @@ open class Pane<T : ComponentView>(
      *
      * @return Coordinate of given child in this container relative to containers anchor point.
      */
-    override fun getChildPosition(child: ComponentView): Coordinate = Coordinate(child.posX, child.posY)
+    override fun getChildPosition(child: ComponentView): Coordinate =
+        Coordinate(child.posX, child.posY)
+    
+    /**
+     * Returning a contained child's coordinates within this container with scale.
+     *
+     * @param child Child to find.
+     *
+     * @return Coordinate of given child in this container relative to containers anchor point.
+     */
+    override fun getActualChildPosition(child: ComponentView): Coordinate =
+        Coordinate(child.actualPosX, child.actualPosY)
     
     /**
      * Removes [component] from container's children.